Use x64 version of DisplaySwitch.exe if available
This will make the "Monitor" remote use the native version of
DisplaySwitch.exe if the user is running a x64 version of
Windows. It seems that Microsoft removed the x86 executable
in one of the recent Windows 10 insider builds (noticed on
build 15014) and thus breaking the monitor output switching
functionality.

You can read up on "File System Redirector", which manifests the
issue here:

https://msdn.microsoft.com/en-us/library/windows/desktop/aa384187.aspx